+++ applications/Unity.GrantManager/modules/Unity.AI/src/Unity.AI.Application/AI/Runtime/IOpenAIResponseParser.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using Unity.AI.Responses;
+
+namespace Unity.AI.Runtime;
+
+internal interface IOpenAIResponseParser
+{
+    ApplicationAnalysisResponse ParseApplicationAnalysisResponse(string raw);
+    ApplicationScoringResponse ParseApplicationScoringResponse(string raw, IReadOnlyDictionary<string, string>? questionIdAliasMap = null);
+}
