+++ applications/Unity.GrantManager/modules/Unity.AI/src/Unity.AI.Application/AI/Runtime/IOpenAITransportService.cs
@@ -1,0 +1,16 @@
+using System.Threading.Tasks;
+using Unity.AI.Models;
+
+namespace Unity.AI.Runtime;
+
+internal interface IOpenAITransportService
+{
+    Task<AIOperationResult> GenerateSummaryAsync(
+        string content,
+        string? systemPrompt,
+        int maxTokens = 150,
+        double? temperature = null,
+        string? operationName = null,
+        string? promptVersion = null,
+        string? fileName = null);
+}
