+++ applications/Unity.GrantManager/modules/Unity.AI/src/Unity.AI.Application/AI/Runtime/OpenAIRuntimeService.cs
@@ -28,6 +28,9 @@ namespace Unity.AI.Runtime
         private readonly IConfiguration _configuration;
         private readonly ILogger<OpenAIRuntimeService> _logger;
         private readonly ITextExtractionService _textExtractionService;
+        private readonly IOpenAITransportService _openAITransportService;
+        private readonly IOpenAIResponseParser _openAIResponseParser;
+        private readonly IOpenAIPromptRenderer _openAIPromptRenderer;
         private readonly ICurrentTenant _currentTenant;
         private readonly IHostEnvironment _hostEnvironment;
         private const string ApplicationAnalysisPromptType = AIPromptTypes.ApplicationAnalysis;
@@ -82,6 +85,9 @@ namespace Unity.AI.Runtime
             IConfiguration configuration,
             ILogger<OpenAIRuntimeService> logger,
             ITextExtractionService textExtractionService,
+            IOpenAITransportService openAITransportService,
+            IOpenAIResponseParser openAIResponseParser,
+            IOpenAIPromptRenderer openAIPromptRenderer,
             ICurrentTenant currentTenant,
             IHostEnvironment hostEnvironment)
         {
@@ -89,6 +95,9 @@ namespace Unity.AI.Runtime
             _configuration = configuration;
             _logger = logger;
             _textExtractionService = textExtractionService;
+            _openAITransportService = openAITransportService;
+            _openAIResponseParser = openAIResponseParser;
+            _openAIPromptRenderer = openAIPromptRenderer;
             _currentTenant = currentTenant;
             _hostEnvironment = hostEnvironment;
         }
@@ -107,7 +116,7 @@ namespace Unity.AI.Runtime
         public async Task<AICompletionResponse> GenerateCompletionAsync(AICompletionRequest request)
         {
             var result = await GenerateWithRetryAsync(
-                () => GenerateSummaryAsync(
+                () => _openAITransportService.GenerateSummaryAsync(
                 request?.UserPrompt ?? string.Empty,
                 null,
                 request?.MaxTokens ?? DefaultCompletionTokens,
@@ -120,7 +129,7 @@ namespace Unity.AI.Runtime
         public async Task<ApplicationAnalysisResponse> GenerateApplicationAnalysisAsync(ApplicationAnalysisRequest request)
         {
             ArgumentNullException.ThrowIfNull(request);
-            var promptVersion = ResolvePromptVersion(request.PromptVersion ?? ResolvePromptVersionSetting(ApplicationAnalysisPromptType));
+            var promptVersion = _openAIPromptRenderer.ResolvePromptVersion(request.PromptVersion ?? ResolvePromptVersionSetting(ApplicationAnalysisPromptType));
             var data = JsonSerializer.Serialize(request.Data, JsonLogOptions);
             var schema = JsonSerializer.Serialize(request.Schema, JsonLogOptions);
 
@@ -133,15 +142,15 @@ namespace Unity.AI.Runtime
                 .Cast<object>();
 
             var attachments = JsonSerializer.Serialize(attachmentsPayload, JsonLogOptions);
-            var systemPrompt = BuildApplicationAnalysisSystemPrompt(promptVersion);
-            var applicationAnalysisContent = BuildApplicationAnalysisUserPrompt(
+            var systemPrompt = _openAIPromptRenderer.BuildApplicationAnalysisSystemPrompt(promptVersion);
+            var applicationAnalysisContent = _openAIPromptRenderer.BuildApplicationAnalysisUserPrompt(
                 promptVersion,
                 schema,
                 data,
                 attachments);
             await LogPromptInputAsync(ApplicationAnalysisPromptType, promptVersion, systemPrompt, applicationAnalysisContent);
             var result = await GenerateWithRetryAsync(
-                () => GenerateSummaryAsync(
+                () => _openAITransportService.GenerateSummaryAsync(
                     applicationAnalysisContent,
                     systemPrompt,
                     ApplicationAnalysisCompletionTokens,
@@ -156,124 +165,7 @@ namespace Unity.AI.Runtime
                 return new ApplicationAnalysisResponse();
             }
 
-            return ParseApplicationAnalysisResponse(AddIdsToAnalysisItems(result.Content));
-        }
-
-        private async Task<AIOperationResult> GenerateSummaryAsync(
-            string content,
-            string? systemPrompt,
-            int maxTokens = 150,
-            double? temperature = null,
-            string? operationName = null,
-            string? promptVersion = null,
-            string? fileName = null)
-        {
-            var providerName = ResolveProviderName(operationName);
-            if (!string.Equals(providerName, DefaultProviderName, StringComparison.Ordinal))
-            {
-                _logger.LogWarning("Provider {ProviderName} is not supported by OpenAIRuntimeService.", providerName);
-                return AIOperationResult.PermanentFailure(new AIProviderResult($"Unsupported provider: {providerName}"));
-            }
-
-            var apiKey = ResolveApiKey(operationName);
-            if (string.IsNullOrEmpty(apiKey))
-            {
-                _logger.LogWarning("Error: {Message}", MissingApiKeyMessage);
-                return AIOperationResult.PermanentFailure(new AIProviderResult(MissingApiKeyMessage));
-            }
-
-            _logger.LogDebug("Calling OpenAI chat completions. PromptLength: {PromptLength}, MaxTokens: {MaxTokens}", content?.Length ?? 0, maxTokens);
-
-            try
-            {
-                var resolvedSystemPrompt = string.IsNullOrWhiteSpace(systemPrompt)
-                    ? "You are a professional grant analyst for the BC Government."
-                    : systemPrompt;
-                var userPrompt = content ?? string.Empty;
-
-                var requestBody = new
-                {
-                    messages = new[]
-                    {
-                       new { role = "system", content = resolvedSystemPrompt },
-                       new { role = "user", content = userPrompt }
-                   }
-                };
-
-                var requestPayload = new Dictionary<string, object?>
-                {
-                    ["messages"] = requestBody.messages,
-                    [ResolveMaxTokensParameterNameForOperation(operationName)] = maxTokens
-                };
-
-                var resolvedTemperature = temperature ?? ResolveConfiguredTemperature(operationName);
-                if (resolvedTemperature.HasValue)
-                {
-                    requestPayload["temperature"] = resolvedTemperature.Value;
-                }
-
-                var json = JsonSerializer.Serialize(requestPayload);
-                var httpContent = new StringContent(json, Encoding.UTF8, "application/json");
-
-                _httpClient.DefaultRequestHeaders.Clear();
-                _httpClient.DefaultRequestHeaders.Add("Authorization", apiKey);
-
-                var response = await _httpClient.PostAsync(ResolveApiUrl(operationName), httpContent);
-                var responseContent = await response.Content.ReadAsStringAsync();
-                var metadata = TryExtractProviderMetadata(responseContent);
-                var providerResponse = BuildProviderResponseFromMetadata(
-                    string.Empty,
-                    responseContent,
-                    metadata,
-                    (int)response.StatusCode);
-
-                _logger.LogDebug(
-                    "OpenAI chat completions response received. StatusCode: {StatusCode}, ResponseLength: {ResponseLength}",
-                    response.StatusCode,
-                    responseContent?.Length ?? 0);
-                LogProviderMetadata(operationName, promptVersion, fileName, providerResponse, response.IsSuccessStatusCode);
-
-                if (!response.IsSuccessStatusCode)
-                {
-                    _logger.LogError("OpenAI API request failed: {StatusCode} - {Content}", response.StatusCode, responseContent);
-                    return MapFailureOutcome(response.StatusCode, providerResponse);
-                }
-
-                if (string.IsNullOrWhiteSpace(responseContent))
-                {
-                    return AIOperationResult.InvalidOutput(providerResponse);
-                }
-
-                try
-                {
-                    using var jsonDoc = JsonDocument.Parse(responseContent);
-                    var choices = jsonDoc.RootElement.GetProperty("choices");
-                    if (choices.GetArrayLength() > 0)
-                    {
-                        var message = choices[0].GetProperty("message");
-                        var modelOutput = message.GetProperty("content").GetString();
-                        return string.IsNullOrWhiteSpace(modelOutput)
-                            ? AIOperationResult.InvalidOutput(providerResponse)
-                            : AIOperationResult.Success(BuildProviderResponseFromMetadata(
-                                modelOutput,
-                                responseContent,
-                                metadata,
-                                (int)response.StatusCode));
-                    }
-
-                    return AIOperationResult.InvalidOutput(providerResponse);
-                }
-                catch (Exception ex) when (ex is JsonException || ex is KeyNotFoundException || ex is InvalidOperationException)
-                {
-                    _logger.LogWarning(ex, "AI response payload had an invalid output shape");
-                    return AIOperationResult.InvalidOutput(providerResponse);
-                }
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "Error generating AI summary");
-                return AIOperationResult.TransientFailure(new AIProviderResult(ex.Message));
-            }
+            return _openAIResponseParser.ParseApplicationAnalysisResponse(result.Content);
         }
 
         public async Task<AttachmentSummaryResponse> GenerateAttachmentSummaryAsync(AttachmentSummaryRequest request)
@@ -282,12 +174,12 @@ namespace Unity.AI.Runtime
             var fileName = request.FileName ?? string.Empty;
             var fileContent = request.FileContent ?? Array.Empty<byte>();
             var contentType = request.ContentType ?? "application/octet-stream";
-            var promptVersion = ResolvePromptVersion(request.PromptVersion ?? ResolvePromptVersionSetting(AttachmentSummaryPromptType));
+            var promptVersion = _openAIPromptRenderer.ResolvePromptVersion(request.PromptVersion ?? ResolvePromptVersionSetting(AttachmentSummaryPromptType));
 
             try
             {
                 var extractedText = await _textExtractionService.ExtractTextAsync(fileName, fileContent, contentType);
-                var prompt = BuildAttachmentSummarySystemPrompt(promptVersion);
+                var prompt = _openAIPromptRenderer.BuildAttachmentSummarySystemPrompt(promptVersion);
 
                 var attachmentText = string.IsNullOrWhiteSpace(extractedText) ? null : extractedText;
                 if (attachmentText != null)
@@ -307,11 +199,11 @@ namespace Unity.AI.Runtime
                     text = attachmentText
                 };
                 var attachment = JsonSerializer.Serialize(attachmentPayload, JsonLogOptions);
-                var contentToAnalyze = BuildAttachmentSummaryUserPrompt(promptVersion, attachment);
+                var contentToAnalyze = _openAIPromptRenderer.BuildAttachmentSummaryUserPrompt(promptVersion, attachment);
 
                 await LogPromptInputAsync(AttachmentSummaryPromptType, promptVersion, prompt, contentToAnalyze);
-                var result = await GenerateWithRetryAsync(
-                () => GenerateSummaryAsync(
+            var result = await GenerateWithRetryAsync(
+                () => _openAITransportService.GenerateSummaryAsync(
                     contentToAnalyze,
                     prompt,
                     AttachmentSummaryCompletionTokens,
@@ -345,81 +237,10 @@ namespace Unity.AI.Runtime
             }
         }
 
-        private string AddIdsToAnalysisItems(string analysisJson)
-        {
-            try
-            {
-                using var jsonDoc = JsonDocument.Parse(analysisJson);
-                using var memoryStream = new System.IO.MemoryStream();
-                using (var writer = new Utf8JsonWriter(memoryStream, new JsonWriterOptions { Indented = true }))
-                {
-                    writer.WriteStartObject();
-
-                    foreach (var property in jsonDoc.RootElement.EnumerateObject())
-                    {
-                        var outputPropertyName = property.Name;
-
-                        if (outputPropertyName == AIJsonKeys.Errors ||
-                            outputPropertyName == AIJsonKeys.Warnings ||
-                            outputPropertyName == AIJsonKeys.Summaries ||
-                            outputPropertyName == AIJsonKeys.Recommendations)
-                        {
-                            writer.WritePropertyName(outputPropertyName);
-                            writer.WriteStartArray();
-
-                            foreach (var item in property.Value.EnumerateArray())
-                            {
-                                writer.WriteStartObject();
-
-                                // Add unique ID first
-                                writer.WriteString("id", Guid.NewGuid().ToString());
-                                writer.WriteBoolean(AIJsonKeys.Dismissed, false);
-
-                                // Copy existing properties
-                                foreach (var itemProperty in item.EnumerateObject())
-                                {
-                                    if (itemProperty.NameEquals(AIJsonKeys.Id) || itemProperty.NameEquals(AIJsonKeys.Dismissed))
-                                    {
-                                        continue;
-                                    }
-
-                                    itemProperty.WriteTo(writer);
-                                }
-
-                                writer.WriteEndObject();
-                            }
-
-                            writer.WriteEndArray();
-                        }
-                        else
-                        {
-                            if (outputPropertyName != property.Name)
-                            {
-                                writer.WritePropertyName(outputPropertyName);
-                                property.Value.WriteTo(writer);
-                                continue;
-                            }
-
-                            property.WriteTo(writer);
-                        }
-                    }
-
-                    writer.WriteEndObject();
-                }
-
-                return Encoding.UTF8.GetString(memoryStream.ToArray());
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "Error adding IDs to analysis items, returning original JSON");
-                return analysisJson; // Return original if processing fails
-            }
-        }
-
         public async Task<ApplicationScoringResponse> GenerateApplicationScoringAsync(ApplicationScoringRequest request)
         {
             ArgumentNullException.ThrowIfNull(request);
-            var promptVersion = ResolvePromptVersion(request.PromptVersion ?? ResolvePromptVersionSetting(ApplicationScoringPromptType));
+            var promptVersion = _openAIPromptRenderer.ResolvePromptVersion(request.PromptVersion ?? ResolvePromptVersionSetting(ApplicationScoringPromptType));
             var dataJson = JsonSerializer.Serialize(request.Data, JsonLogOptions);
             var sectionJson = JsonSerializer.Serialize(request.SectionSchema, JsonLogOptions);
 
@@ -438,8 +259,8 @@ namespace Unity.AI.Runtime
                     ? string.Join("\n- ", attachmentSummaries.Select((summary, index) => $"Attachment {index + 1}: {summary}"))
                     : "No attachments provided.";
 
-                var section = BuildAliasedApplicationScoringSection(request.SectionName, sectionJson, out var questionIdAliasMap);
-                var response = BuildApplicationScoringResponseTemplate(section);
+                var section = _openAIPromptRenderer.BuildAliasedApplicationScoringSection(request.SectionName, sectionJson, out var questionIdAliasMap);
+                var response = _openAIPromptRenderer.BuildApplicationScoringResponseTemplate(section);
                 if (response == "{}")
                 {
                     _logger.LogWarning(
@@ -448,17 +269,17 @@ namespace Unity.AI.Runtime
                     return new ApplicationScoringResponse();
                 }
 
-                var applicationScoringContent = BuildApplicationScoringUserPrompt(
+                var applicationScoringContent = _openAIPromptRenderer.BuildApplicationScoringUserPrompt(
                     promptVersion,
                     dataJson,
                     attachments,
                     section,
                     response);
-                var systemPrompt = BuildApplicationScoringSystemPrompt(promptVersion);
+                var systemPrompt = _openAIPromptRenderer.BuildApplicationScoringSystemPrompt(promptVersion);
 
                 await LogPromptInputAsync(ApplicationScoringPromptType, promptVersion, systemPrompt, applicationScoringContent);
                 var result = await GenerateWithRetryAsync(
-                () => GenerateSummaryAsync(
+                () => _openAITransportService.GenerateSummaryAsync(
                     applicationScoringContent,
                     systemPrompt,
                     ApplicationScoringCompletionTokens,
@@ -473,7 +294,7 @@ namespace Unity.AI.Runtime
                     return new ApplicationScoringResponse();
                 }
 
-                return ParseApplicationScoringResponse(result.Content, questionIdAliasMap);
+                return _openAIResponseParser.ParseApplicationScoringResponse(result.Content, questionIdAliasMap);
             }
             catch (Exception ex)
             {
@@ -865,259 +686,6 @@ namespace Unity.AI.Runtime
             return trimmedEndpoint + trimmedPath;
         }
 
-        private static ApplicationAnalysisResponse ParseApplicationAnalysisResponse(string raw)
-        {
-            var response = new ApplicationAnalysisResponse();
-
-            if (!TryParseJsonObjectFromResponse(raw, out var root))
-            {
-                return response;
-            }
-
-            if (TryGetStringProperty(root, AIJsonKeys.Decision, out var decision))
-            {
-                response.Decision = decision?.Trim().ToUpperInvariant();
-            }
-
-            if (root.TryGetProperty("errors", out var errors) && errors.ValueKind == JsonValueKind.Array)
-            {
-                response.Errors = ParseFindings(errors);
-            }
-
-            if (root.TryGetProperty("warnings", out var warnings) && warnings.ValueKind == JsonValueKind.Array)
-            {
-                response.Warnings = ParseFindings(warnings);
-            }
-
-            if (root.TryGetProperty(AIJsonKeys.Summaries, out var summaries) && summaries.ValueKind == JsonValueKind.Array)
-            {
-                response.Summaries = ParseFindings(summaries);
-            }
-
-            if (root.TryGetProperty(AIJsonKeys.Recommendations, out var recommendations) && recommendations.ValueKind == JsonValueKind.Array)
-            {
-                response.Recommendations = ParseFindings(recommendations);
-            }
-
-            return response;
-        }
-
-        private static bool TryGetStringProperty(JsonElement root, string propertyName, out string? value)
-        {
-            value = null;
-            if (!root.TryGetProperty(propertyName, out var property) || property.ValueKind != JsonValueKind.String)
-            {
-                return false;
-            }
-
-            value = property.GetString();
-            return !string.IsNullOrWhiteSpace(value);
-        }
-
-        private static List<ApplicationAnalysisFinding> ParseFindings(JsonElement array)
-        {
-            var findings = new List<ApplicationAnalysisFinding>();
-            foreach (var item in array.EnumerateArray())
-            {
-                if (item.ValueKind != JsonValueKind.Object)
-                {
-                    continue;
-                }
-
-                var id = item.TryGetProperty(AIJsonKeys.Id, out var idProp) && idProp.ValueKind == JsonValueKind.String
-                    ? idProp.GetString()
-                    : null;
-                var dismissed = item.TryGetProperty(AIJsonKeys.Dismissed, out var dismissedProp) &&
-                    (dismissedProp.ValueKind == JsonValueKind.True || dismissedProp.ValueKind == JsonValueKind.False) &&
-                    dismissedProp.GetBoolean();
-                string? title = null;
-                if (item.TryGetProperty(AIJsonKeys.Title, out var titleProp) && titleProp.ValueKind == JsonValueKind.String)
-                {
-                    title = titleProp.GetString();
-                }
-
-                string? detail = null;
-                if (item.TryGetProperty(AIJsonKeys.Detail, out var detailProp) && detailProp.ValueKind == JsonValueKind.String)
-                {
-                    detail = detailProp.GetString();
-                }
-
-                findings.Add(new ApplicationAnalysisFinding
-                {
-                    Id = id,
-                    Dismissed = dismissed,
-                    Title = title,
-                    Detail = detail
-                });
-            }
-
-            return findings;
-        }
-
-        private static ApplicationScoringResponse ParseApplicationScoringResponse(
-            string raw,
-            IReadOnlyDictionary<string, string>? questionIdAliasMap = null)
-        {
-            var response = new ApplicationScoringResponse();
-            if (!TryParseJsonObjectFromResponse(raw, out var root))
-            {
-                return response;
-            }
-
-            foreach (var property in root.EnumerateObject())
-            {
-                if (property.Value.ValueKind != JsonValueKind.Object)
-                {
-                    continue;
-                }
-
-                var answer = property.Value.TryGetProperty("answer", out var answerProp)
-                    ? answerProp.Clone()
-                    : default;
-                var rationale = property.Value.TryGetProperty("rationale", out var rationaleProp) &&
-                                rationaleProp.ValueKind == JsonValueKind.String
-                    ? rationaleProp.GetString() ?? string.Empty
-                    : string.Empty;
-                var confidence = property.Value.TryGetProperty("confidence", out var confidenceProp) &&
-                                 confidenceProp.ValueKind == JsonValueKind.Number &&
-                                 confidenceProp.TryGetInt32(out var parsedConfidence)
-                    ? NormalizeConfidence(parsedConfidence)
-                    : 0;
-
-                var questionId = questionIdAliasMap != null &&
-                                 questionIdAliasMap.TryGetValue(property.Name, out var originalQuestionId)
-                    ? originalQuestionId
-                    : property.Name;
-
-                response.Answers[questionId] = new ApplicationScoringAnswer
-                {
-                    Answer = answer,
-                    Rationale = rationale,
-                    Confidence = confidence
-                };
-            }
-
-            return response;
-        }
-
-        private static int NormalizeConfidence(int confidence)
-        {
-            var clamped = Math.Clamp(confidence, 0, 100);
-            var rounded = (int)Math.Round(clamped / 5.0, MidpointRounding.AwayFromZero) * 5;
-            return Math.Clamp(rounded, 0, 100);
-        }
-
-        private static string BuildApplicationScoringResponseTemplate(string sectionPayloadJson)
-        {
-            try
-            {
-                using var doc = JsonDocument.Parse(sectionPayloadJson);
-                if (!doc.RootElement.TryGetProperty("questions", out var questions) || questions.ValueKind != JsonValueKind.Array)
-                {
-                    return "{}";
-                }
-
-                var template = new Dictionary<string, object>();
-                foreach (var question in questions.EnumerateArray())
-                {
-                    if (!question.TryGetProperty("id", out var idProp) || idProp.ValueKind != JsonValueKind.String)
-                    {
-                        continue;
-                    }
-
-                    var questionId = idProp.GetString();
-                    if (string.IsNullOrWhiteSpace(questionId))
-                    {
-                        continue;
-                    }
-
-                    template[questionId] = new
-                    {
-                        answer = string.Empty,
-                        rationale = string.Empty,
-                        confidence = 0
-                    };
-                }
-
-                if (template.Count == 0)
-                {
-                    return "{}";
-                }
-
-                return JsonSerializer.Serialize(template, JsonLogOptions);
-            }
-            catch (JsonException)
-            {
-                return "{}";
-            }
-        }
-
-        private static string BuildAliasedApplicationScoringSection(
-            string? sectionName,
-            string sectionJson,
-            out IReadOnlyDictionary<string, string> questionIdAliasMap)
-        {
-            questionIdAliasMap = new Dictionary<string, string>(StringComparer.Ordinal);
-
-            if (string.IsNullOrWhiteSpace(sectionJson))
-            {
-                return JsonSerializer.Serialize(new { name = sectionName, questions = sectionJson }, JsonLogOptions);
-            }
-
-            try
-            {
-                using var sectionDoc = JsonDocument.Parse(sectionJson);
-                if (sectionDoc.RootElement.ValueKind != JsonValueKind.Array)
-                {
-                    return JsonSerializer.Serialize(new { name = sectionName, questions = sectionDoc.RootElement.Clone() }, JsonLogOptions);
-                }
-
-                var aliasedQuestions = new List<Dictionary<string, object?>>();
-                var aliasMap = new Dictionary<string, string>(StringComparer.Ordinal);
-                var index = 1;
-
-                foreach (var question in sectionDoc.RootElement.EnumerateArray())
-                {
-                    if (question.ValueKind != JsonValueKind.Object)
-                    {
-                        continue;
-                    }
-
-                    var aliasedQuestion = new Dictionary<string, object?>(StringComparer.Ordinal);
-                    string? questionAlias = null;
-
-                    foreach (var property in question.EnumerateObject())
-                    {
-                        if (property.NameEquals("id") && property.Value.ValueKind == JsonValueKind.String)
-                        {
-                            var originalQuestionId = property.Value.GetString();
-                            if (!string.IsNullOrWhiteSpace(originalQuestionId))
-                            {
-                                questionAlias = $"q{index++}";
-                                aliasMap[questionAlias] = originalQuestionId;
-                                aliasedQuestion[property.Name] = questionAlias;
-                                continue;
-                            }
-                        }
-
-                        aliasedQuestion[property.Name] = property.Value.Clone();
-                    }
-
-                    if (!string.IsNullOrWhiteSpace(questionAlias))
-                    {
-                        aliasedQuestions.Add(aliasedQuestion);
-                    }
-                }
-
-                questionIdAliasMap = aliasMap;
-                return JsonSerializer.Serialize(new { name = sectionName, questions = aliasedQuestions }, JsonLogOptions);
-            }
-            catch (JsonException)
-            {
-                return JsonSerializer.Serialize(new { name = sectionName, questions = sectionJson }, JsonLogOptions);
-            }
-        }
-
         private async Task LogPromptInputAsync(string promptType, string promptVersion, string? systemPrompt, string userPrompt)
         {
             var formattedInput = FormatPromptInputForLog(systemPrompt, userPrompt);
@@ -1300,255 +868,6 @@ namespace Unity.AI.Runtime
             {
                 return false;
             }
-        }
-
-        private static string ResolvePromptVersion(string? version)
-        {
-            if (!string.IsNullOrWhiteSpace(version) &&
-                PromptProfiles.TryGetValue(version.Trim(), out var selectedVersion))
-            {
-                return selectedVersion;
-            }
-
-            return PromptVersionV1;
-        }
-
-        private static string BuildApplicationAnalysisSystemPrompt(string version)
-        {
-            return GetRequiredPromptTemplate(version, ApplicationAnalysisSystemTemplateName);
-        }
-
-        private static string BuildApplicationAnalysisUserPrompt(
-            string version,
-            string schema,
-            string data,
-            string attachments)
-        {
-            var replacements = new Dictionary<string, string>
-            {
-                ["SCHEMA"] = schema,
-                ["DATA"] = data,
-                ["ATTACHMENTS"] = attachments
-            };
-
-            return RenderPromptTemplate(version, ApplicationAnalysisUserTemplateName, replacements);
-        }
-
-        private static string BuildAttachmentSummarySystemPrompt(string version)
-        {
-            return GetRequiredPromptTemplate(version, AttachmentSummarySystemTemplateName);
-        }
-
-        private static string BuildAttachmentSummaryUserPrompt(string version, string attachment)
-        {
-            return RenderPromptTemplate(version, AttachmentSummaryUserTemplateName, new Dictionary<string, string>
-            {
-                ["ATTACHMENT"] = attachment
-            });
-        }
-
-        private static string BuildApplicationScoringSystemPrompt(string version)
-        {
-            return GetRequiredPromptTemplate(version, ApplicationScoringSystemTemplateName);
-        }
-
-        private static string BuildApplicationScoringUserPrompt(
-            string version,
-            string data,
-            string attachments,
-            string section,
-            string response)
-        {
-            return RenderPromptTemplate(version, ApplicationScoringUserTemplateName, new Dictionary<string, string>
-            {
-                ["DATA"] = data,
-                ["ATTACHMENTS"] = attachments,
-                ["SECTION"] = section,
-                ["RESPONSE"] = response
-            });
-        }
-
-        private static bool TryGetPromptTemplate(string version, string templateName, out string template)
-        {
-            template = string.Empty;
-            var cacheKey = $"{version}/{templateName}";
-            if (PromptTemplateCache.TryGetValue(cacheKey, out var cachedTemplate))
-            {
-                template = cachedTemplate;
-                return true;
-            }
-
-            var path = Path.Combine(AppContext.BaseDirectory, PromptTemplatesFolder, version, $"{templateName}.txt");
-            if (!File.Exists(path))
-            {
-                return false;
-            }
-
-            var loaded = PromptTemplateCache.GetOrAdd(cacheKey, _ => File.ReadAllText(path));
-            if (string.IsNullOrWhiteSpace(loaded))
-            {
-                return false;
-            }
-
-            template = loaded;
-            return true;
-        }
-
-        private static string GetRequiredPromptTemplate(string version, string templateName)
-        {
-            if (TryGetPromptTemplate(version, templateName, out var template))
-            {
-                return template;
-            }
-
-            throw new InvalidOperationException(
-                $"Missing required prompt template '{templateName}.txt' for prompt version '{version}'.");
-        }
-
-        private static string RenderPromptTemplate(
-            string version,
-            string templateName,
-            IReadOnlyDictionary<string, string> runtimeReplacements)
-        {
-            return RenderPromptTemplateInternal(
-                version,
-                templateName,
-                runtimeReplacements,
-                new HashSet<string>(StringComparer.OrdinalIgnoreCase));
-        }
-
-        private static string RenderPromptTemplateInternal(
-            string version,
-            string templateName,
-            IReadOnlyDictionary<string, string> runtimeReplacements,
-            ISet<string> resolutionStack)
-        {
-            if (!resolutionStack.Add(templateName))
-            {
-                throw new InvalidOperationException(
-                    $"Detected cyclic prompt fragment reference while resolving '{templateName}.txt' for prompt version '{version}'.");
-            }
-
-            var template = GetRequiredPromptTemplate(version, templateName);
-            var replacements = new Dictionary<string, string>(runtimeReplacements, StringComparer.Ordinal);
-            var baseTemplateName = GetTemplateBaseName(templateName);
-
-            foreach (var placeholder in GetTemplatePlaceholders(template))
-            {
-                if (replacements.ContainsKey(placeholder))
-                {
-                    continue;
-                }
-
-                var fragmentTemplateName = ResolveFragmentTemplateName(version, baseTemplateName, placeholder);
-                if (!string.IsNullOrWhiteSpace(fragmentTemplateName))
-                {
-                    replacements[placeholder] = RenderPromptTemplateInternal(
-                        version,
-                        fragmentTemplateName,
-                        new Dictionary<string, string>(StringComparer.Ordinal),
-                        resolutionStack).TrimEnd();
-                }
-            }
-
-            var rendered = template;
-            foreach (var replacement in replacements)
-            {
-                rendered = rendered.Replace($"{{{{{replacement.Key}}}}}", replacement.Value ?? string.Empty, StringComparison.Ordinal);
-            }
-
-            var unresolved = GetTemplatePlaceholders(rendered);
-            if (unresolved.Count > 0)
-            {
-                throw new InvalidOperationException(
-                    $"Unresolved prompt placeholders in '{templateName}.txt' for prompt version '{version}': {string.Join(", ", unresolved.OrderBy(item => item))}");
-            }
-
-            resolutionStack.Remove(templateName);
-            return rendered;
-        }
-
-        private static string? ResolveFragmentTemplateName(string version, string baseTemplateName, string placeholderName)
-        {
-            var normalizedPlaceholder = placeholderName.ToLowerInvariant();
-            var baseScopedCandidate = $"{baseTemplateName}.{normalizedPlaceholder}";
-            if (TryGetPromptTemplate(version, baseScopedCandidate, out _))
-            {
-                return baseScopedCandidate;
-            }
-
-            if (string.Equals(placeholderName, "RESPONSE", StringComparison.Ordinal))
-            {
-                var outputCandidate = $"{baseTemplateName}.output";
-                if (TryGetPromptTemplate(version, outputCandidate, out _))
-                {
-                    return outputCandidate;
-                }
-            }
-
-            if (TryResolveCommonTemplateName(placeholderName, out var commonTemplateName) &&
-                TryGetPromptTemplate(version, commonTemplateName, out _))
-            {
-                return commonTemplateName;
-            }
-
-            return null;
-        }
-
-        private static bool TryResolveCommonTemplateName(string placeholderName, out string commonTemplateName)
-        {
-            commonTemplateName = string.Empty;
-            if (!placeholderName.StartsWith("COMMON_", StringComparison.Ordinal))
-            {
-                return false;
-            }
-
-            var suffix = placeholderName.Substring("COMMON_".Length).ToLowerInvariant();
-            suffix = suffix.Replace('_', '.');
-            commonTemplateName = $"common.{suffix}";
-            return true;
-        }
-
-        private static string GetTemplateBaseName(string templateName)
-        {
-            var separatorIndex = templateName.IndexOf('.', StringComparison.Ordinal);
-            if (separatorIndex <= 0)
-            {
-                return templateName;
-            }
-
-            return templateName.Substring(0, separatorIndex);
-        }
-
-        private static HashSet<string> GetTemplatePlaceholders(string template)
-        {
-            var placeholders = new HashSet<string>(StringComparer.Ordinal);
-            var searchIndex = 0;
-
-            while (searchIndex < template.Length)
-            {
-                var start = template.IndexOf("{{", searchIndex, StringComparison.Ordinal);
-                if (start < 0)
-                {
-                    break;
-                }
-
-                var end = template.IndexOf("}}", start + 2, StringComparison.Ordinal);
-                if (end < 0)
-                {
-                    break;
-                }
-
-                var placeholder = template.Substring(start + 2, end - start - 2).Trim();
-                if (!string.IsNullOrWhiteSpace(placeholder))
-                {
-                    placeholders.Add(placeholder);
-                }
-
-                searchIndex = end + 2;
-            }
-
-            return placeholders;
         }
 
         private static string ExtractSummaryFromJson(string output)
